Move applyFetchHooks to element hooks module 🎨
